Rename module to "bagger"
Preparing for release/merging with ReBACH
