+++ .idea/misc.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10.6 WSL (Ubuntu-22.04): (/home/jnr/code/ReDATA-preservation/.venv/bin/python)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="/home/jnr/code/ReDATA-preservation/.venv/bin/python" project-jdk-type="Python SDK" />
 </project>
