+++ bagger/config/default_workflow.json
