+++ bagger/job.py
