+++ bagger/metadata.py
