+++ bagger/strip.py
